+++ crates/protocol/src/platform.rs
@@ -12,29 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use alloc::borrow::Cow;
+
 use wasefire_common::platform::Side;
 use wasefire_error::Error;
 use wasefire_wire::Wire;
 
 #[derive(Debug, Wire)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Info<'a> {
-    pub serial: &'a [u8],
+    pub serial: Cow<'a, [u8]>,
     pub running_side: Side,
-    pub running_version: &'a [u8],
-    pub opposite_version: Result<&'a [u8], Error>,
+    pub running_version: Cow<'a, [u8]>,
+    pub opposite_version: Result<Cow<'a, [u8]>, Error>,
 }
 
 #[cfg(feature = "host")]
 impl core::fmt::Display for Info<'_> {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         use data_encoding::HEXLOWER as HEX;
-        writeln!(f, "  serial: {}", HEX.encode_display(self.serial))?;
+        writeln!(f, "  serial: {}", HEX.encode_display(&self.serial))?;
         writeln!(f, "    side: {}", self.running_side)?;
-        writeln!(f, " version: {}", HEX.encode_display(self.running_version))?;
+        writeln!(f, " version: {}", HEX.encode_display(&self.running_version))?;
         writeln!(
             f,
             "opposite: {}",
-            match self.opposite_version {
+            match &self.opposite_version {
                 Ok(x) => HEX.encode(x),
                 Err(e) => alloc::format!("{e}"),
             }
@@ -43,7 +46,8 @@ impl core::fmt::Display for Info<'_> {
 }
 
 #[derive(Debug, Wire)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct _Info0<'a> {
-    pub serial: &'a [u8],
-    pub version: &'a [u8],
+    pub serial: Cow<'a, [u8]>,
+    pub version: Cow<'a, [u8]>,
 }