+++ crates/protocol/test.sh
@@ -21,6 +21,6 @@ test_helper
 
 cargo test --lib --features=_descriptor,host
 cargo test --lib --features=_descriptor,device
+cargo test --lib --features=serde
 cargo check --lib --features=host
 cargo check --lib --target=thumbv7em-none-eabi --features=device
-cargo check --lib --features=serde